pyproject.toml
[project]
name = "pam-gmm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4.0",
    "django-import-export>=4.1.1",
    "django-unfold>=0.43.0",
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "gunicorn>=23.0.0",
    "matplotlib>=3.9.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "plotly>=5.24.1",
    "redis>=5.2.0",
    "seaborn>=0.13.2",
    "whitenoise>=6.8.2",
    "kaleido==0.2.1",
]
